requirements.txt
connexion[swagger-ui]
flask-mongoengine
flask-bcrypt
flask-restful